requirements.txt
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
facebook-sdk==3.1.0
feedparser==5.2.1
future==0.18.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
oauthlib==3.1.0
pycodestyle==2.5.0
pylint==2.4.4
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
